% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout.R
\name{breakout}
\alias{breakout}
\title{Function to breakout region in gcamdata}
\usage{
breakout(gcamdataFolder = NULL, countryNew = NULL)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}

\item{countryNew}{Default = NULL. Name of New Country.}
}
\description{
Function to breakout region in gcamdata
}
