% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout_city.R
\name{breakout_city}
\alias{breakout_city}
\title{Function to breakout region in gcamdata}
\usage{
breakout_city(
  gcamdataFolder = NULL,
  region = NULL,
  city = NULL,
  popProjection = NULL,
  pcgdpProjection = NULL
)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}

\item{region}{Default = NULL. Name of The region from which to break out the city.}

\item{city}{Default = NULL. Name of city to breakout.}

\item{popProjection}{Default = NULL. Projection of population for city.}

\item{pcgdpProjection}{Default = NULL. Projection of per capita gdp for city.}
}
\description{
Function to breakout region in gcamdata
}
