% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout_regions.R
\name{breakout_regions}
\alias{breakout_regions}
\title{Function to breakout region in gcamdata}
\usage{
breakout_regions(
  gcamdataFolder = NULL,
  regionsNew = NULL,
  countriesNew = NULL,
  gcam_version = "6.0"
)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}

\item{regionsNew}{Default = NULL. Name of New Region or regions to breakout.
Can be a vector or list. (e.g. regionsNew = "New Region" OR regionsNew = c("New Region 1","New Region 2"))}

\item{countriesNew}{Default = NULL. Name of countries in new region.
A vector if only one region or a list with countries for each regionsNew.
(e.g. c("New Region Country 1, "New Region Country 2) corresponding to a single regionsNew = "New Region" OR
 countriesNew = list(c("New Region 1 Country 1", "New Region 1 Country 2"), c("New Region 2 Country 1", "New Region 2 Country 2"))
 corresponding to regionsNew = c("New Region 1", "New Region 2")))}

\item{gcam_version}{Default = 6.0. Which version of GCAM are the changes being applied to. Default may work on multiple}
}
\description{
Function to breakout region in gcamdata
}
