% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout_subregion.R
\name{breakout_subregion}
\alias{breakout_subregion}
\title{Function to breakout subregion in gcamdata}
\usage{
breakout_subregion(
  gcamdataFolder = NULL,
  region = NULL,
  subregion = NULL,
  pop_projection = NULL,
  pcgdp_projection = NULL
)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}

\item{region}{Default = NULL. Name of The region from which to break out the subregion.}

\item{subregion}{Default = NULL. Name of subregion to breakout.}

\item{pop_projection}{Default = NULL. Projection of population for subregion.}

\item{pcgdp_projection}{Default = NULL. Projection of per capita gdp for subregion.}
}
\description{
Function to breakout subregion in gcamdata
}
