% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout_subregions.R
\name{breakout_subregions}
\alias{breakout_subregions}
\title{Function to breakout subregion in gcamdata}
\usage{
breakout_subregions(
  gcamdataFolder = NULL,
  region = NULL,
  pop_projection = NULL,
  pcgdp_projection = NULL,
  industry_shares = NULL,
  gcam_version = "6.0"
)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}

\item{region}{Default = NULL. Name of The region from which to break out the subregion.}

\item{pop_projection}{Default = NULL. Projection of population (millions) for subregions and rest of region.}

\item{pcgdp_projection}{Default = NULL. Projection of per capita gdp (thous 2005 USD per capita) for subregions and rest of region.}

\item{gcam_version}{Default = 6.0. Which version of GCAM are the changes being applied to. Default may work on multiple}
}
\description{
Function to breakout subregion in gcamdata
}
