% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakoutRegion.R
\name{breakoutRegion}
\alias{breakoutRegion}
\title{Function to breakout region in gcamdata}
\usage{
breakoutRegion(gcamdataFolder = NULL, regionNew = NULL, countriesNew = NULL)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}

\item{regionNew}{Default = NULL. Name of New Region to breakout.}

\item{countriesNew}{Default = NULL. Name of countries in new region.}
}
\description{
Function to breakout region in gcamdata
}
