% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcambreakout.R
\docType{package}
\name{gcambreakout}
\alias{gcambreakout}
\title{gcambreakout}
\description{
\itemize{
\item Github: https://github.com/JGCRI/gcambreakout
\item Webpage: https://jgcri.github.io/gcambreakout/}
}
