% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IEA_EnergyBalances_Countries_2019}
\alias{IEA_EnergyBalances_Countries_2019}
\title{IEA_EnergyBalances_Countries_2019}
\format{
.csv
}
\source{
IEA_EnergyBalances_2019.csv.gz
}
\usage{
IEA_EnergyBalances_Countries_2019
}
\description{
IEA_EnergyBalances_Countries_2019
}
\keyword{datasets}
