% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_modules}
\alias{mapping_modules}
\title{Mapping of modules in gcamdata system which give errors for certain countries.}
\format{
R dataframe
}
\source{
created by ZK
}
\usage{
mapping_modules
}
\description{
Mapping of modules in gcamdata system which give errors for certain countries.
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::mapping_modules
}
}
\keyword{datasets}
