% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\title{Function to breakout region in gcamdata}
\usage{
restore(gcamdataFolder = NULL)
}
\arguments{
\item{gcamdataFolder}{Default = NULL. Full path to gcamdata folder.}
}
\description{
Function to breakout region in gcamdata
}
