% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcambreakout.R
\docType{package}
\name{rgcambreakout}
\alias{rgcambreakout}
\title{rgcambreakout}
\description{
\itemize{
\item Github: https://github.com/JGCRI/rgcambreakout
\item Webpage: https://jgcri.github.io/rgcambreakout/}
}
