% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_breakout_helpers}
\alias{template_breakout_helpers}
\title{Template for breakout-helpers.R}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_breakout_helpers
}
\description{
Template for breakout-helpers.R
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_breakout_helpers
}
}
\keyword{datasets}
