% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_IEA_memo_ctry_comments}
\alias{template_IEA_memo_ctry_comments}
\title{template_IEA_memo_ctry_comments}
\format{
.csv
}
\source{
template_IEA_memo_ctry_comments
}
\usage{
template_IEA_memo_ctry_comments
}
\description{
template_IEA_memo_ctry_comments
}
\keyword{datasets}
