% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_industry_shares}
\alias{template_industry_shares}
\title{Template for industryShares}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_industry_shares
}
\description{
Template for industryShares
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_industry_shares
}
}
\keyword{datasets}
