% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_pcgdpProjection}
\alias{template_pcgdpProjection}
\title{Template for pcgdpProjection file inputs}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_pcgdpProjection
}
\description{
Template for pcgdpProjection file inputs
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_pcgdpProjection
}
}
\keyword{datasets}
