% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_pop_projection}
\alias{template_pop_projection}
\title{Template for popProjection}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_pop_projection
}
\description{
Template for popProjection
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_pop_projection
}
}
\keyword{datasets}
