% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_popProjection}
\alias{template_popProjection}
\title{Template for popProjection}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_popProjection
}
\description{
Template for popProjection
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_popProjection
}
}
\keyword{datasets}
