% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zchunk_L2231.wind_update_breakout}
\alias{template_zchunk_L2231.wind_update_breakout}
\title{Template for template_zchunk_L2231.wind_update_breakout}
\format{
R script
}
\source{
created by ZK
}
\usage{
template_zchunk_L2231.wind_update_breakout
}
\description{
Template for template_zchunk_L2231.wind_update_breakout
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_zchunk_L2231.wind_update_breakout
}
}
\keyword{datasets}
