% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zchunk_LB142.ag_Fert_IO_R_C_Y_GLU_breakout_gcamv6p0}
\alias{template_zchunk_LB142.ag_Fert_IO_R_C_Y_GLU_breakout_gcamv6p0}
\title{Template for template_zchunk_LB142.ag_Fert_IO_R_C_Y_GLU_breakout_gcamv6p0}
\format{
R script
}
\source{
created by BY
}
\usage{
template_zchunk_LB142.ag_Fert_IO_R_C_Y_GLU_breakout_gcamv6p0
}
\description{
Template for template_zchunk_LB142.ag_Fert_IO_R_C_Y_GLU_breakout_gcamv6p0
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_zchunk_LB142.ag_Fert_IO_R_C_Y_GLU_breakout_gcamv6p0
}
}
\keyword{datasets}
