% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zchunk_X201.socioeconomic_APPEND}
\alias{template_zchunk_X201.socioeconomic_APPEND}
\title{Template for zchunk_X201.socioeconomic_APPEND.R}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_zchunk_X201.socioeconomic_APPEND
}
\description{
Template for zchunk_X201.socioeconomic_APPEND.R
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_zchunk_X201.socioeconomic_APPEND
}
}
\keyword{datasets}
