% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zchunk_X254.transportation_APPEND_gcam6p0}
\alias{template_zchunk_X254.transportation_APPEND_gcam6p0}
\title{Template for zchunk_X254.transportation_APPEND_gcam6p0.R}
\format{
R script
}
\source{
created by NB, ZK
}
\usage{
template_zchunk_X254.transportation_APPEND_gcam6p0
}
\description{
Template for zchunk_X254.transportation_APPEND_gcam6p0.R
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::zchunk_X254.transportation_APPEND_gcam6p0
}
}
\keyword{datasets}
