% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zemissions_L112.ceds_ghg_en_R_S_T_Y_breakout_gcamv7p0}
\alias{template_zemissions_L112.ceds_ghg_en_R_S_T_Y_breakout_gcamv7p0}
\title{template_zemissions_L112.ceds_ghg_en_R_S_T_Y_breakout_gcamv7p0}
\format{
R script
}
\source{
created by ZK
}
\usage{
template_zemissions_L112.ceds_ghg_en_R_S_T_Y_breakout_gcamv7p0
}
\description{
template_zemissions_L112.ceds_ghg_en_R_S_T_Y_breakout_gcamv7p0
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_zemissions_L112.ceds_ghg_en_R_S_T_Y_breakout_gcamv7p0
}
}
\keyword{datasets}
