% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zenergy_L120.offshore_wind_breakout_gcamv7p0}
\alias{template_zenergy_L120.offshore_wind_breakout_gcamv7p0}
\title{Template for template_zenergy_L120.offshore_wind_breakout_gcamv7p0}
\format{
R script
}
\source{
created by ZK
}
\usage{
template_zenergy_L120.offshore_wind_breakout_gcamv7p0
}
\description{
Template for template_zenergy_L120.offshore_wind_breakout_gcamv7p0
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_zenergy_L120.offshore_wind_breakout_gcamv7p0
}
}
\keyword{datasets}
