% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_zsocio_L2323.iron_steel_Inc_Elas_scenarios_breakout_gcamv7p0}
\alias{template_zsocio_L2323.iron_steel_Inc_Elas_scenarios_breakout_gcamv7p0}
\title{template_zsocio_L2323.iron_steel_Inc_Elas_scenarios_breakout_gcamv7p0}
\format{
R script
}
\source{
created by ZK
}
\usage{
template_zsocio_L2323.iron_steel_Inc_Elas_scenarios_breakout_gcamv7p0
}
\description{
template_zsocio_L2323.iron_steel_Inc_Elas_scenarios_breakout_gcamv7p0
}
\examples{
\dontrun{
 library(gcambreakout);
 gcambreakout::template_zsocio_L2323.iron_steel_Inc_Elas_scenarios_breakout_gcamv7p0
}
}
\keyword{datasets}
